Added comments and code style changes
